chore: moved desktop folder in electron
