+++ guardoni/src/electron/OutputPanel.tsx
