+++ guardoni/src/electron/config.ts
